Move IPP api models to payments.inperson package
