+++ example/src/test/java/org/wordpress/android/fluxc/wc/pay/WCInPersonPaymentsStoreTest.kt
@@ -9,8 +9,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.junit.MockitoJUnitRunner
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooPayload
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.pay.ConnectionTokenApiResponse
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.pay.WCInPersonPaymentsRestClient
+import org.wordpress.android.fluxc.network.rest.wpcom.wc.payments.inperson.ConnectionTokenApiResponse
+import org.wordpress.android.fluxc.network.rest.wpcom.wc.payments.inperson.WCInPersonPaymentsRestClient
 import org.wordpress.android.fluxc.store.WCInPersonPaymentsStore
 import org.wordpress.android.fluxc.test
 import org.wordpress.android.fluxc.tools.initCoroutineEngine
