+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/network/rest/wpcom/wc/payments/inperson/ConnectionTokenApiResponse.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.fluxc.network.rest.wpcom.wc.pay
+package org.wordpress.android.fluxc.network.rest.wpcom.wc.payments.inperson
 
 import com.google.gson.annotations.SerializedName
